# 49_Se7enSquare-main/back-end/docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/communities": {
      "get": {
        "operationId": "CommunitiesController_findAll",
        "summary": "Get all communities",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role for RBAC: admin | moderator | user",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CommunityDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "communities"
        ]
      },
      "post": {
        "operationId": "CommunitiesController_create",
        "summary": "Create community",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role for RBAC: admin | moderator | user",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCommunityDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommunityDto"
                }
              }
            }
          }
        },
        "tags": [
          "communities"
        ]
      }
    },
    "/communities/{id}": {
      "get": {
        "operationId": "CommunitiesController_findOne",
        "summary": "Get community by id",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role for RBAC: admin | moderator | user",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommunityDto"
                }
              }
            }
          }
        },
        "tags": [
          "communities"
        ]
      },
      "patch": {
        "operationId": "CommunitiesController_update",
        "summary": "Update community",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role for RBAC: admin | moderator | user",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCommunityDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommunityDto"
                }
              }
            }
          }
        },
        "tags": [
          "communities"
        ]
      },
      "delete": {
        "operationId": "CommunitiesController_remove",
        "summary": "Delete community",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role for RBAC: admin | moderator | user",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "Community 3 deleted"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "communities"
        ]
      }
    },
    "/events": {
      "get": {
        "operationId": "EventsController_findAll",
        "summary": "Get all events",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role for RBAC: admin | moderator | user",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EventDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "events"
        ]
      },
      "post": {
        "operationId": "EventsController_create",
        "summary": "Create event",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role for RBAC: admin | moderator | user",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEventDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventDto"
                }
              }
            }
          }
        },
        "tags": [
          "events"
        ]
      }
    },
    "/events/{id}": {
      "get": {
        "operationId": "EventsController_findOne",
        "summary": "Get event by id",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role for RBAC: admin | moderator | user",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventDto"
                }
              }
            }
          }
        },
        "tags": [
          "events"
        ]
      },
      "patch": {
        "operationId": "EventsController_update",
        "summary": "Update event",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role for RBAC: admin | moderator | user",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEventDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventDto"
                }
              }
            }
          }
        },
        "tags": [
          "events"
        ]
      },
      "delete": {
        "operationId": "EventsController_remove",
        "summary": "Delete event",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role for RBAC: admin | moderator | user",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "Event 2 deleted"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "events"
        ]
      }
    },
    "/reports": {
      "get": {
        "operationId": "ReportsController_findAll",
        "summary": "List reports",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role for RBAC: admin | moderator | user",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReportDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "reports"
        ]
      },
      "post": {
        "operationId": "ReportsController_create",
        "summary": "Create report",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role for RBAC: admin | moderator | user",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReportDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportDto"
                }
              }
            }
          }
        },
        "tags": [
          "reports"
        ]
      }
    },
    "/reports/{id}/status": {
      "patch": {
        "operationId": "ReportsController_updateStatus",
        "summary": "Update report status (pending -> reviewed -> resolved)",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role for RBAC: admin | moderator | user",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateReportStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportDto"
                }
              }
            }
          }
        },
        "tags": [
          "reports"
        ]
      }
    },
...
  },
  "info": {
    "title": "Se7enSquare Backend API",
    "description": "In-memory NestJS backend for frontend integration and academic evaluation",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "x-role": {
        "type": "apiKey",
        "in": "header",
        "name": "x-role",
        "description": "RBAC role header: admin | moderator | user"
      }
    },
    "schemas": {
...
      "CommunityDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "FPS Arena"
          },
          "description": {
            "type": "string",
            "example": "Competitive FPS players and tournaments"
          },
          "ownerId": {
            "type": "number",
            "example": 3
          },
          "tags": {
            "example": [
              "fps",
              "esports"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "ownerId",
          "tags"
        ]
      },
      "CreateCommunityDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 3,
            "maxLength": 50,
            "example": "MOBA Masters"
          },
          "description": {
            "type": "string",
            "minLength": 10,
            "maxLength": 200,
            "example": "Community for strategy and ranked discussions"
          },
          "ownerId": {
            "type": "number",
            "example": 3
          },
          "tags": {
            "example": [
              "moba",
              "ranked"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "name",
          "description",
          "ownerId",
          "tags"
        ]
      },
      "UpdateCommunityDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 3,
            "maxLength": 50,
            "example": "MOBA Masters"
          },
          "description": {
            "type": "string",
            "minLength": 10,
            "maxLength": 200,
            "example": "Community for strategy and ranked discussions"
          },
          "ownerId": {
            "type": "number",
            "example": 3
          },
          "tags": {
            "example": [
              "moba",
              "ranked"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "EventDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1
          },
          "title": {
            "type": "string",
            "example": "Friday Scrim Night"
          },
          "description": {
            "type": "string",
            "example": "Weekly custom matches"
          },
          "communityId": {
            "type": "number",
            "example": 1
          },
          "date": {
            "type": "string",
            "example": "2026-05-02T18:00:00.000Z"
          }
        },
        "required": [
          "id",
          "title",
          "description",
          "communityId",
          "date"
        ]
      },
      "CreateEventDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "minLength": 3,
            "maxLength": 60,
            "example": "Clan Tournament Qualifier"
          },
          "description": {
            "type": "string",
            "minLength": 10,
            "maxLength": 220,
            "example": "Qualifier round for monthly tournament"
          },
          "communityId": {
            "type": "number",
            "example": 1
          },
          "date": {
            "type": "string",
            "example": "2026-05-20T16:30:00.000Z"
          }
        },
        "required": [
          "title",
          "description",
          "communityId",
          "date"
        ]
      },
      "UpdateEventDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "minLength": 3,
            "maxLength": 60,
            "example": "Clan Tournament Qualifier"
          },
          "description": {
            "type": "string",
            "minLength": 10,
            "maxLength": 220,
            "example": "Qualifier round for monthly tournament"
          },
          "communityId": {
            "type": "number",
            "example": 1
          },
          "date": {
            "type": "string",
            "example": "2026-05-20T16:30:00.000Z"
          }
        }
      },
      "ReportDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1
          },
          "reporterId": {
            "type": "number",
            "example": 3
          },
          "targetType": {
            "type": "string",
            "enum": [
              "post",
              "user",
              "community"
            ],
            "example": "post"
          },
          "targetId": {
            "type": "number",
            "example": 1
          },
          "reason": {
            "type": "string",
            "example": "Potential harassment in comments"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "reviewed",
              "resolved"
            ],
            "example": "pending"
          }
        },
        "required": [
          "id",
          "reporterId",
          "targetType",
          "targetId",
          "reason",
          "status"
        ]
      },
      "CreateReportDto": {
        "type": "object",
        "properties": {
          "reporterId": {
            "type": "number",
            "example": 3
          },
          "targetType": {
            "type": "string",
            "enum": [
              "post",
              "user",
              "community"
            ],
            "example": "post"
          },
          "targetId": {
            "type": "number",
            "example": 1
          },
          "reason": {
            "type": "string",
            "minLength": 8,
            "maxLength": 200,
            "example": "Abusive language found in comments"
          }
        },
        "required": [
          "reporterId",
          "targetType",
          "targetId",
          "reason"
        ]
      },
      "UpdateReportStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "reviewed",
              "resolved"
            ],
            "example": "reviewed"
          }
        },
        "required": [
          "status"
        ]
      },
...
    }
  }
}